Move remaning tests to the tests/ dir
