+++ portal/common/lib/tests/domain_parsing.suffix_mock.test.ts
@@ -3,7 +3,7 @@
 
 import { describe, expect, test } from 'vitest'
 import { vi, beforeEach } from 'vitest'
-import { splitUrl } from './domain_parsing'
+import { splitUrl } from '../src/domain_parsing'
 
 /**
  * Make sure that splitURL supports both cases where `wal.app` is both included and not included
